# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Employee Service",
        "contact": {
            "name": "DEVils"
        },
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/employee": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Полуение конкретного сотрудника",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cutlery47_employee-service_internal_model.GetEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cutlery47_employee-service_internal_model.GetEmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/employees": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Полуение сотрудников по фильтрам",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cutlery47_employee-service_internal_model.GetBaseEmployeesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cutlery47_employee-service_internal_model.BaseEmployee"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "github_com_cutlery47_employee-service_internal_model.BaseEmployee": {
            "type": "object",
            "properties": {
                "family_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_general": {
                    "type": "boolean"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "github_com_cutlery47_employee-service_internal_model.GetBaseEmployeesRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "github_com_cutlery47_employee-service_internal_model.GetEmployeeRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "github_com_cutlery47_employee-service_internal_model.GetEmployeeResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "family_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office_address": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "teammates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cutlery47_employee-service_internal_model.BaseEmployee"
                    }
                }
            }
        }
    }
}